Bug fix and creation of new GPT.ini file
DSC resource was having trouble due to a bug in the cmdlets.  If an array containing a single UInt32 value is passed to -Data when setting a DWord value (or a single UInt64 value and a QWord), the command should work now (and was throwing an error before).

Also, the commands will now create a default GPT.ini file, if one doesn't exist and the -NoUpdateGptIni switch is not used.