Signing change
PolFileEditor.dll is signed in source control and doesn't change.  I've modified the build script so that after the next release, deployment scripts can be written to just check the hash of module files and only overwrite the ones that it needs to.

(This is coming up due to "file in use" errors when trying to delete or overwrite the DLL file, since the LCM still has it loaded sometimes.)