Error reporting changes
The code is less lazy about reporting errors now.  It builds ErrorRecord objects with module-specific FullyQualifiedErrorId values, and the top-level (exported) functions use $PSCmdlet.ThrowTerminatingError() to return those values to the caller.  This results in errors that look more like what you get from a cmdlet, and should be more useful to users of the module.

Since this could technically be a breaking change for scripts that are looking for the old errors, I'll be bumping the version number up to v2.0.

I don't have a lot of tests around the error conditions at the moment; should add some later to make sure that predictable errors produce the output that they should.